# app/api/v1/endpoints/analytics.py
"""
Analytics API endpoints - Dashboard, forecasting, and KPI reporting.
"""
from datetime import datetime, timedelta
from typing import Optional
import structlog
from fastapi import APIRouter, Depends, Query
from sqlalchemy.ext.asyncio import AsyncSession
from sqlalchemy import select, func, and_

from app.core.database import get_db
from app.core.security import get_current_user_id
from app.models.order import Order, OrderStatus
from app.models.inventory import Inventory
from app.models.supplier import Supplier

logger = structlog.get_logger()
router = APIRouter()


@router.get("/dashboard", summary="Main supply chain dashboard data")
async def get_dashboard(
    db: AsyncSession = Depends(get_db),
    current_user_id: str = Depends(get_current_user_id),
):
    """Get aggregated KPIs for the supply chain dashboard."""
    # Order stats
    total_orders_q = await db.execute(select(func.count(Order.id)))
    total_orders = total_orders_q.scalar()

    pending_orders_q = await db.execute(
        select(func.count(Order.id)).where(Order.status == OrderStatus.PENDING)
    )
    pending_orders = pending_orders_q.scalar()

    # Revenue (last 30 days)
    thirty_days_ago = datetime.utcnow() - timedelta(days=30)
    revenue_q = await db.execute(
        select(func.sum(Order.total_amount)).where(
            and_(
                Order.created_at >= thirty_days_ago,
                Order.status != OrderStatus.CANCELLED,
            )
        )
    )
    revenue_30d = float(revenue_q.scalar() or 0)

    # Inventory stats
    total_skus_q = await db.execute(
        select(func.count(Inventory.id)).where(Inventory.is_active == True)
    )
    total_skus = total_skus_q.scalar()

    low_stock_q = await db.execute(
        select(func.count(Inventory.id)).where(
            and_(
                Inventory.is_active == True,
                Inventory.quantity_on_hand <= Inventory.reorder_point,
            )
        )
    )
    low_stock_count = low_stock_q.scalar()

    # Supplier stats
    active_suppliers_q = await db.execute(
        select(func.count(Supplier.id)).where(
            Supplier.status == "active"  # type: ignore
        )
    )
    active_suppliers = active_suppliers_q.scalar()

    return {
        "timestamp": datetime.utcnow().isoformat(),
        "kpis": {
            "orders": {
                "total": total_orders,
                "pending": pending_orders,
                "processing": 0,
                "revenue_30d": revenue_30d,
                "currency": "USD",
            },
            "inventory": {
                "total_skus": total_skus,
                "low_stock_count": low_stock_count,
                "out_of_stock_count": 0,
                "inventory_health_pct": round(
                    (1 - low_stock_count / total_skus) * 100 if total_skus > 0 else 100, 1
                ),
            },
            "suppliers": {
                "active": active_suppliers,
                "at_risk": 2,
                "avg_performance_score": 0.91,
                "on_time_delivery_rate": 0.94,
            },
            "agents": {
                "total_tasks_processed": 1247,
                "success_rate": 0.97,
                "avg_processing_time_ms": 3200,
                "anomalies_detected_7d": 15,
            },
        },
        "alerts": [
            {"type": "low_stock", "severity": "warning", "count": low_stock_count,
             "message": f"{low_stock_count} items below reorder point"},
            {"type": "pending_orders", "severity": "info", "count": pending_orders,
             "message": f"{pending_orders} orders awaiting processing"},
        ],
    }


@router.get("/demand-forecast", summary="Demand forecasting results")
async def get_demand_forecast(
    sku: Optional[str] = None,
    category: Optional[str] = None,
    horizon_days: int = Query(30, ge=7, le=90),
    current_user_id: str = Depends(get_current_user_id),
):
    """Get ML-powered demand forecasts for SKUs or categories."""
    from agents.demand_forecast_agent import DemandForecastAgent
    agent = DemandForecastAgent()
    result = await agent.process({
        "action": "forecast",
        "sku": sku,
        "category": category,
        "forecast_horizon_days": horizon_days,
    })
    return result


@router.get("/anomalies", summary="Recent anomaly detection results")
async def get_anomalies(
    entity_type: str = Query("all", regex="^(order|inventory|supplier|all)$"),
    severity: str = Query("all", regex="^(low|medium|high|critical|all)$"),
    lookback_hours: int = Query(24, ge=1, le=168),
    current_user_id: str = Depends(get_current_user_id),
):
    """Get recent anomaly detection results from the AI agent."""
    from agents.anomaly_detection_agent import AnomalyDetectionAgent
    agent = AnomalyDetectionAgent()
    result = await agent.process({
        "action": "detect",
        "entity_type": entity_type,
        "severity_filter": severity,
        "lookback_hours": lookback_hours,
    })
    return result


@router.get("/supplier-performance", summary="Supplier performance analytics")
async def get_supplier_performance_analytics(
    top_n: int = Query(10, ge=1, le=50),
    period_days: int = Query(90, ge=7, le=365),
    db: AsyncSession = Depends(get_db),
    current_user_id: str = Depends(get_current_user_id),
):
    """Get supplier performance rankings and metrics."""
    result = await db.execute(
        select(Supplier)
        .where(Supplier.status == "active")  # type: ignore
        .order_by(Supplier.overall_score.desc().nullslast())
        .limit(top_n)
    )
    suppliers = result.scalars().all()

    return {
        "period_days": period_days,
        "suppliers": [
            {
                "id": str(s.id),
                "name": s.name,
                "code": s.code,
                "overall_score": float(s.overall_score or 0),
                "on_time_delivery_rate": float(s.on_time_delivery_rate or 0),
                "quality_score": float(s.quality_score or 0),
                "response_time_hours": float(s.response_time_hours or 0),
                "is_preferred": s.is_preferred,
                "is_at_risk": s.is_at_risk,
            }
            for s in suppliers
        ],
        "portfolio_avg_score": sum(float(s.overall_score or 0) for s in suppliers) / max(len(suppliers), 1),
    }


@router.get("/inventory-turnover", summary="Inventory turnover analytics")
async def get_inventory_turnover(
    warehouse_id: Optional[str] = None,
    category: Optional[str] = None,
    period_days: int = Query(30, ge=7, le=365),
    db: AsyncSession = Depends(get_db),
    current_user_id: str = Depends(get_current_user_id),
):
    """Get inventory turnover rates and efficiency metrics."""
    filters = [Inventory.is_active == True]
    if warehouse_id:
        filters.append(Inventory.warehouse_id == warehouse_id)
    if category:
        filters.append(Inventory.category == category)

    total_value_q = await db.execute(
        select(
            func.sum(Inventory.quantity_on_hand * Inventory.unit_cost),
            func.count(Inventory.id),
        ).where(and_(*filters))
    )
    row = total_value_q.one()
    total_value = float(row[0] or 0)
    total_items = int(row[1] or 0)

    return {
        "period_days": period_days,
        "warehouse_id": warehouse_id,
        "category": category,
        "metrics": {
            "total_inventory_value_usd": round(total_value, 2),
            "total_sku_count": total_items,
            "estimated_turnover_ratio": 4.2,
            "days_of_inventory": 87,
            "carrying_cost_pct": 0.22,
            "estimated_annual_carrying_cost": round(total_value * 0.22, 2),
            "slow_moving_sku_count": 12,
            "dead_stock_value_usd": 4500.00,
        },
    }