# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Supply Chain AI Platform API
  description: |
    Production-grade multi-agent AI supply chain orchestration API powered by Claude Opus 4.6.

    ## Features
    - **Multi-Agent Orchestration**: 7 specialized AI agents coordinated by Claude Opus 4.6
    - **Real-time Inventory Management**: Stock tracking with automated reorder recommendations
    - **Order Processing**: End-to-end order lifecycle with fraud detection
    - **Supplier Management**: Vendor evaluation and RFQ generation
    - **Demand Forecasting**: ML-based predictions using XGBoost
    - **Anomaly Detection**: Real-time fraud and anomaly scoring

    ## Authentication
    All endpoints require JWT Bearer token authentication (except `/health`, `/ready`, and `/api/v1/auth/token`).

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    Obtain tokens via `POST /api/v1/auth/token` (OAuth2 password flow).

    ## Rate Limiting
    - 100 requests/minute per authenticated user
    - 1000 requests/hour per authenticated user
    - Agent orchestration: 10 requests/minute per user
  version: "1.0.0"
  contact:
    name: Platform Team
    email: platform@supply-chain.internal
  license:
    name: Proprietary
    url: https://supply-chain.internal/license
servers:
  - url: http://localhost:8000
    description: Local Development
  - url: https://api.supply-chain.internal
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/auth/token:
    post:
      tags: [Authentication]
      summary: Get access token
      description: OAuth2 password flow. Returns JWT access and refresh tokens.
      security: []
      operationId: login_for_access_token
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin@supply-chain.local
                password:
                  type: string
                  format: password
      responses:
        "200":
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      operationId: register_user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: User created successfully
        "409":
          description: Email or username already exists
...
  /api/v1/agents/orchestrate:
    post:
      tags: [Agents]
      summary: Orchestrate a multi-agent task
      operationId: orchestrate_task
      description: |
        Submit a task for multi-agent orchestration. Returns task_id for polling.
        The orchestrator agent will route to appropriate specialized agents.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrchestrationRequest"
      responses:
        "202":
          description: Task accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrchestrationResponse"
  /api/v1/agents/tasks/{task_id}:
    get:
      tags: [Agents]
      summary: Get task status
      operationId: get_task_status
      parameters:
        - name: task_id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Task status and result
...
  /api/v1/analytics/demand-forecast:
    get:
      tags: [Analytics]
      summary: Get demand forecast
      operationId: get_demand_forecast
      parameters:
        - name: product_id
          in: query
          schema: {type: string}
        - name: horizon_days
          in: query
          schema: {type: integer, default: 30}
      responses:
        "200":
          description: Demand forecast with confidence intervals
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
...
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          example: 1800
    UserCreate:
      type: object
      required: [email, username, password]
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 8
        full_name:
          type: string
...
    OrchestrationRequest:
      type: object
      required: [task_description]
      properties:
        task_description:
          type: string
          maxLength: 5000
        task_type:
          type: string
          enum: [process_order, check_inventory, evaluate_supplier, route_shipment, analyze_demand, detect_anomaly, general]
          default: general
        priority:
          type: string
          enum: [low, normal, high, critical]
          default: normal
        context:
          type: object
          description: Additional context for the agent
        metadata:
          type: object
    OrchestrationResponse:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
          example: processing
        estimated_completion_seconds:
          type: integer
        poll_url:
          type: string